+++ dsort/dsort_mem.go
@@ -350,7 +350,7 @@ func (ds *dsorterMem) createShardsLocally() (err error) {
 	metrics.begin()
 	defer metrics.finish()
 	metrics.Lock()
-	metrics.ToCreate = len(phaseInfo.metadata.Shards)
+	metrics.ToCreate = int64(len(phaseInfo.metadata.Shards))
 	metrics.Unlock()
 
 	var (
